+++ src/main/java/org/iesvdm/modelo/ComercialDTO.java
@@ -3,6 +3,8 @@ package org.iesvdm.modelo;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class ComercialDTO implements Comparable<ComercialDTO>{
@@ -11,8 +13,7 @@ public class ComercialDTO implements Comparable<ComercialDTO>{
     private String nombre;
     private String apellido1;
     private String apellido2;
-    private float comision;
-
+    private BigDecimal comision;
 
     private int pedidosPorCliente;
